+++ run.sh
@@ -1,2 +1,2 @@
 source install/setup.sh
-ros2 launch frontier_explorer sim_slam_bringup.launch.py 
+ros2 launch autonomousr_explorer_bringup full_system.launch.py